split some linkable components of ft_0CDD
sdata2 says these must be separate
