+++ config/GALE01/splits.txt
@@ -999,9 +999,17 @@ melee/ft/ftlipstickswing.c:
 	.sdata2     start:0x804D8F10 end:0x804D8F14
 
 melee/ft/ft_0CDD.c:
-	.text       start:0x800CDD14 end:0x800CECE8
+	.text       start:0x800CDD14 end:0x800CDF5C
 	.sdata      start:0x804D3CC0 end:0x804D3CD0
-	.sdata2     start:0x804D8F18 end:0x804D8F30
+	.sdata2     start:0x804D8F18 end:0x804D8F20
+
+melee/ft/ft_0CDF.c:
+	.text       start:0x800CDF5C end:0x800CE3E0
+	.sdata2     start:0x804D8F20 end:0x804D8F28
+
+melee/ft/ft_0CE3.c:
+	.text       start:0x800CE3E0 end:0x800CECE8
+	.sdata2     start:0x804D8F28 end:0x804D8F30
 
 melee/ft/ftattacks4combo.c:
 	.text       start:0x800CECE8 end:0x800CEE70
